more application restructuring
