+++ env.template
@@ -14,9 +14,3 @@ MESSAGE_SUBJECT=flair
 LOGGING=True
 # Set frequency at which bot accepts commands (typically 1)
 REFRESH_INTERVAL=1
-
-
-# DEPRECATED
-GOLD_LIMIT=1489626000
-CURRENT_SEASON=35
-REFRESH_INTERVAL=1
